+++ ground/src/main/java/com/google/android/ground/ui/datacollection/tasks/point/DropPinTaskMapFragment.kt
@@ -16,11 +16,8 @@
 package com.google.android.ground.ui.datacollection.tasks.point
 
 import android.os.Bundle
-import androidx.hilt.navigation.fragment.hiltNavGraphViewModels
-import com.google.android.ground.R
 import com.google.android.ground.ui.common.AbstractMapFragmentWithControls
 import com.google.android.ground.ui.common.BaseMapViewModel
-import com.google.android.ground.ui.datacollection.DataCollectionViewModel
 import com.google.android.ground.ui.map.CameraPosition
 import com.google.android.ground.ui.map.MapFragment
 import dagger.hilt.android.AndroidEntryPoint
@@ -30,13 +27,10 @@ import javax.inject.Inject
 class DropPinTaskMapFragment @Inject constructor() : AbstractMapFragmentWithControls() {
 
   private lateinit var mapViewModel: BaseMapViewModel
-  private val dataCollectionViewModel: DataCollectionViewModel by
-    hiltNavGraphViewModels(R.id.data_collection)
   private val viewModel: DropPinTaskViewModel by lazy {
     // Access to this viewModel is lazy for testing. This is because the NavHostController could
     // not be initialized before the Fragment under test is created, leading to
     // hiltNavGraphViewModels() to fail when called on launch.
-    val taskId = arguments?.getString(TASK_ID_FRAGMENT_ARG_KEY) ?: error("null taskId fragment arg")
     dataCollectionViewModel.getTaskViewModel(taskId) as DropPinTaskViewModel
   }
 
@@ -55,9 +49,5 @@ class DropPinTaskMapFragment @Inject constructor() : AbstractMapFragmentWithCont
   override fun onMapCameraMoved(position: CameraPosition) {
     super.onMapCameraMoved(position)
     viewModel.updateCameraPosition(position)
-  }
-
-  companion object {
-    const val TASK_ID_FRAGMENT_ARG_KEY = "taskId"
   }
 }
