+++ ground/src/main/java/com/google/android/ground/ui/datacollection/tasks/polygon/DrawAreaTaskMapFragment.kt
@@ -16,12 +16,9 @@
 package com.google.android.ground.ui.datacollection.tasks.polygon
 
 import android.os.Bundle
-import androidx.hilt.navigation.fragment.hiltNavGraphViewModels
 import androidx.lifecycle.lifecycleScope
-import com.google.android.ground.R
 import com.google.android.ground.ui.common.AbstractMapFragmentWithControls
 import com.google.android.ground.ui.common.BaseMapViewModel
-import com.google.android.ground.ui.datacollection.DataCollectionViewModel
 import com.google.android.ground.ui.map.CameraPosition
 import com.google.android.ground.ui.map.Feature
 import com.google.android.ground.ui.map.MapFragment
@@ -33,13 +30,11 @@ import kotlinx.coroutines.launch
 class DrawAreaTaskMapFragment @Inject constructor() : AbstractMapFragmentWithControls() {
 
   private lateinit var mapViewModel: BaseMapViewModel
-  private val dataCollectionViewModel: DataCollectionViewModel by
-    hiltNavGraphViewModels(R.id.data_collection)
+
   private val viewModel: DrawAreaTaskViewModel by lazy {
     // Access to this viewModel is lazy for testing. This is because the NavHostController could
     // not be initialized before the Fragment under test is created, leading to
     // hiltNavGraphViewModels() to fail when called on launch.
-    val taskId = arguments?.getString(TASK_ID_FRAGMENT_ARG_KEY) ?: error("null taskId fragment arg")
     dataCollectionViewModel.getTaskViewModel(taskId) as DrawAreaTaskViewModel
   }
 
@@ -68,9 +63,5 @@ class DrawAreaTaskMapFragment @Inject constructor() : AbstractMapFragmentWithCon
         map.getDistanceInPixels(c1, c2)
       }
     }
-  }
-
-  companion object {
-    const val TASK_ID_FRAGMENT_ARG_KEY = "taskId"
   }
 }
